examples/requirements.txt
perplexity-correlations
filelock
PyYAML
pandas
transformers
datasets
argparse
lm-eval